refactor: consolidate Tempo into Cadence
The two-app split (pure Cadence vs. task-aware Tempo) was a hedge
against the question of whether the task layer was a separate
product. The answer was no — one user, one tool. Empty activity
pools already render the timer without any activity label, so
"pure mode" is one settings toggle away rather than a separate
binary.

Changes:
- Folds TempoApp.swift into Cadence's AppDelegate. ~30 lines of
  activity-related hooks (editor window, phase-completion polling)
  added; the rest of the delegate was already line-for-line
  identical between the two apps.
- Moves TempoTimerView and ActivityEditorView into Sources/Cadence/.
  TempoTimerView renamed to TimerView (Cadence target).
- Drops the unused TimerView composite from CadenceKit; its
  sub-components (PhaseLabel, TimeDisplay, CadenceButton,
  PhaseTimeline, GradientBackground, GrainOverlay) remain since
  the consolidated TimerView still composes them.
- Removes Tempo target and TempoTests target from Package.swift.
- Moves ActivityStoreTests into Tests/CadenceKitTests/ (the test
  was already @testable import CadenceKit).
- Renames the ActivityStore persistence directory from
  ~/Library/Application Support/Tempo/ to .../Cadence/.
- Removes scripts/dev-tempo.sh and scripts/bundle-tempo.sh.
