+++ Sources/CadenceKit/TimerView.swift
@@ -262,55 +262,3 @@ public struct PhaseTimeline: View {
     }
 }
 
-// MARK: - Timer View
-
-public struct TimerView: View {
-    @Bindable var timerState: TimerState
-
-    public init(timerState: TimerState) { self.timerState = timerState }
-
-    public var body: some View {
-        ZStack {
-            GradientBackground(phase: timerState.currentPhase)
-            GrainOverlay()
-
-            VStack {
-                HStack(alignment: .center) {
-                    PhaseLabel(phase: timerState.currentPhase)
-                    Spacer()
-                }
-                .padding(.top, DesignSystem.Spacing.headerTop)
-                .padding(.horizontal, DesignSystem.Spacing.headerSideInset)
-                Spacer()
-            }
-
-            VStack {
-                HStack {
-                    Spacer()
-                    TimeDisplay(secondsRemaining: timerState.secondsRemaining)
-                        .padding(.trailing, DesignSystem.Spacing.timeRightPadding)
-                }
-                .padding(.leading, DesignSystem.Spacing.timeLeftInset)
-                .padding(.top, DesignSystem.Spacing.timeTop)
-                Spacer()
-            }
-
-            VStack(spacing: 0) {
-                Spacer()
-                CadenceButton(
-                    isRunning: timerState.isRunning,
-                    phase: timerState.currentPhase
-                ) { timerState.toggle() }
-                .padding(.bottom, DesignSystem.Spacing.buttonBottom)
-
-                PhaseTimeline(timerState: timerState)
-                    .padding(.horizontal, DesignSystem.Spacing.timelineSideInset)
-                    .padding(.bottom, DesignSystem.Spacing.timelineBottom)
-            }
-        }
-        .frame(
-            width: DesignSystem.Spacing.windowWidth,
-            height: DesignSystem.Spacing.windowHeight
-        )
-    }
-}
